Added TypeScript outline for constants.
