--- src/constants.js
@@ -1,7 +1,0 @@
-export const COLORS = {
-    purple: '#7E62CF',
-    green: '#537A72',
-    peach: '#E5AD83',
-    cream: '#E6CFAC',
-    pink: '#F1778D'
-}
