+++ src/constants.tsx
@@ -1,0 +1,16 @@
+
+interface COLORS {
+  purple: string;
+  green: string;
+  peach: string;
+  cream: string;
+  pink: string;
+}
+
+export const COLORS = {
+  purple: '#7E62CF',
+  green: '#537A72',
+  peach: '#E5AD83',
+  cream: '#E6CFAC',
+  pink: '#F1778D'
+}
